Start rewriting without selenium
