+++ Pipfile
@@ -7,6 +7,8 @@ name = "pypi"
 selenium = "~=4.9.0"
 webdriver-manager = "~=3.8.6"
 keyring = "~=23.13.1"
+requests = "~=2.30.0"
+beautifulsoup4 = "~=4.12.2"
 
 [dev-packages]
